+++ src/main/java/com/example/repository/RepositoryRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface RepositoryRepository  extends JpaRepository<Repository, Long> {
-    List<Repository> findByUsuarioId(Long usuarioId);
+    List<Repository> findByUsuarioId(Integer usuarioId);
 }
